+++ src/extension_core/extension.ts
@@ -110,8 +110,8 @@ export async function activate(context: vscode.ExtensionContext) {
   }));
 
   context.subscriptions.push(vscode.commands.registerCommand('vaporview.removeSignal', (e) => {
-    if (e && e.netlistId !== undefined) {
-      viewerProvider.removeSignalFromDocument(e.netlistId, true);
+    if (e && e.rowId !== undefined) {
+      viewerProvider.removeSignalFromDocument(undefined, e.rowId, true);
     }
   }));